# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/scripts/": {
      "post": {
        "description": "Подставляет наличие и цену из файла источника в целевые файлы",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "multFiles",
            "in": "formData",
            "type": "array",
            "required": true,
            "description": "Загрузите сначала файл sales_dynamics, затем Шаблон загрузки линз",
            "collectionFormat": "multi",
            "items": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/ymProfitGetPersent/": {
      "post": {
        "description": "YM маржа, парсинг процента комиссии. Логин, пароль от аккаунта ЯМ. Результат - файл с столбцами sku, name, persent, delivery.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "password",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "assortimentLink",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "startPage",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "numberOfPage",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "assortimentLink": {
                  "example": "any"
                },
                "startPage": {
                  "example": "any"
                },
                "numberOfPage": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}